split transforms tests
